+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `rails local_credentials:edit` command to manipulate local credentials in Rails 7.1+. ([@palkan][])
+
 ## 2.5.4 (2023-10-15)
 
 - Fix tracing with empty config keys. ([@palkan][])